requirements.txt
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
langchain-google-genai
streamlit
pypdf
docx2txt
faiss-cpu
unstructured[all]
networkx
msoffcrypto-tool
openpyxl
jq
selenium
pdfminer.six
unstructured[image]
pi-heif
pillow
pytesseract
hf_xet
pymongo
motor
langchain-text-splitters
sentence-transformers
gpt4all 
google-generativeai
langchain-huggingface
duckduckgo-search
ddgs
youtube_search
arxiv
Langgraph
langchain-groq
langsmith
pyowm
langchain-ollama


